Moved tests into test_topology.py.
